core/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cite2fn"
version = "0.1.0"
description = "Convert hyperlinked citations in Word documents to Bluebook or APA footnotes, endnotes, or a list of references."
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "python-docx>=1.1.0",
    "lxml>=5.0",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "pymupdf>=1.24",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
cite2fn = "cite2fn.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["cite2fn*"]
